updated namechecker
